+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 Change log
 ================================================================================
 
+0.0.5 - 07.06.2020
+--------------------------------------------------------------------------------
+
+**Updated**
+
+#. Name the moban actions when this template is used
+#. `#12 <https://github.com/moremoban/moban-handlebars/issues/12>`_: provide
+   apis for end user to use their own block helper and partials
+
 0.0.4 - 14.10.2019
 --------------------------------------------------------------------------------